arm: mvebu: theadorable: Add 'pcie' test command
This board specific command tests for the presence of a specified PCIe
device (via vendor-ID and device-ID). If the device is not detected,
this will get printed. If the device is detected, the board will get
resetted so that an easy loop test can be done. The board will reboot
until the PCIe device is not detected.

Signed-off-by: Stefan Roese <sr@denx.de>
(cherry picked from commit 1ec2a80b10c5755cdadbe3c128b607f434e77372)