MLK-18289-2 imx8mm_evk: Add codes to setup typec PD charging
On iMX8MM EVK, the USB2 port is the primary power supply, when USB2 is charging,
the 3483_EN_SNK1 is not enabled. In the board, we init the tcpc of USB2 first with
PD enabled. If it is charging, when init the tcpc of USB1 we disable the PD,
otherwise we enable PD on USB1.

The USB PD switch NX20P3438 needs to exit dead battery mode before
enable EN_SINK, otherwise the OVP is fixed at 6.8V. Also the OVP threshold needs set to
23V, when we switch to use 20V VBUS input.

Due to HW issue, after we switched to 9V/20V, set on/off button to off then set back it to
on, the board can't power up. In this patch we limit the voltage to 5V, will change back
to 9V/20V when HW fixes the issue.

Signed-off-by: Ye Li <ye.li@nxp.com>
Acked-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit e3579303f0e050aeeee938584492f76840e4bb97)