MLK-18787 mx6sxsabreauto: Fix vga stdout and stderr
The u-boot changed to check CONFIG_IS_ENABLED for SYS_CONSOLE_IS_IN_ENV,
so we have to set the configuration in defconfig not in header file,
otherwise the checking will fail and lead to use vga for stdout and stderr
when video is enabled.

Fix the issue by moving CONFIG_VIDEO and CONFIG_SYS_CONSOLE_IS_IN_ENV to
defconfig and remove duplicated configurations from header file.

Signed-off-by: Ye Li <ye.li@nxp.com>