MLK-19623: change hdmi keep-alive check mechanism
The current keep-alive check mechanism uses a static variable that is
initialized to 0. When the function is first called, it may happen to
catch the 8-bit keep-alive counter right when it overflows, hence
returning BUSY.

This patch will keep checking the counter for 10us, every 1us, but it
will immediately return if the keep-alive counter changed.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>