MLK-21988 imx8mn: rom_api: Change to support secondary boot on SD/eMMC
ROM supports secondary boot which can boot from a new offset specified
by fuse. So adjust the offset calculation by counting the ROM's image_offset
parameter for SD/eMMC.
The secondary boot on flexspi is different like SD/eMMC. ROM will map the
space of secondary boot to the start of flexspi AHB memory. So it is same
as primary boot for SPL.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>