MLK-22005 imx8: clock: Handle FEC with freq limited on 8UX
When FEC is fused to limit RMII 10/100, we should configure clocks for RMII.
It needs enable 50Mhz ref clock, set SC_C_TXCLK mux to dummy. And we don't need
to enable the LPCG for RGMII TX CLK for RMII.

The latest SCFW has fixed a issue that PERCLK slice was disabled before
on parts with freq fused. With that issue the LPCGs using PERCLK as sources
are always in STOP staus. So the u-boot will loops in checking STOP status.

"commit 59e7cf4e81dd265411e4f1dcd5186567d3fb51ef
 Author: Ranjani Vaidyanathan <ranjani.vaidyanathan@nxp.com>
 Date:   Mon Jun 17 17:10:43 2019 -0500

    SCF-420 Ensure IEEE1588 can work on fused QX parts
"

SO to work with both old and latest SCFW versions, this patch changes to enable
LPCG by LPCG_ClockOn which won't wait on the LPCG STOP status

Signed-off-by: Ye Li <ye.li@nxp.com>
Acked-by: Peng Fan <peng.fan@nxp.com>